uvcvideo: Avoid cyclic entity chains due to malformed USB descriptors
Way back in 2017, fuzzing the 4.14-rc2 USB stack with syzkaller kicked
up the following WARNING from the UVC chain scanning code:
......

The problem with this descriptor is that it is self-referential: the
source ID of 0 matches itself! This causes the 'struct uvc_entity'
representing the display to be added to its chain list twice during
'uvc_scan_chain()': once via 'uvc_scan_chain_entity()' when it is
processed directly from the 'dev->entities' list and then again
immediately afterwards when trying to follow the source ID in
'uvc_scan_chain_forward()'

Add a check before adding an entity to a chain list to ensure that the
entity is not already part of a chain.

android_id: A-111893654
patch_level: 2020-09-05
CVE: CVE-2020-0404
@type CVE @id: CVE-2020-0404

Change-Id: I3a417fec4050aef42e06a4c4ffc7154c84e80a08
Reviewed-by: soft1-codecheck
Reviewed-by: google-patch-id <xiao.qihu@zte.com.cn>
Tested-by: google-patch-id <xiao.qihu@zte.com.cn>
Autosubmitted-by: google-patch-id <xiao.qihu@zte.com.cn>
Presubmit-Ready-by: google-patch-id <xiao.qihu@zte.com.cn>
Presubmit-Verified-by: google-patch-id <xiao.qihu@zte.com.cn>